src/python/rss/requirements.txt
feedparser
python-dateutil
psycopg2-binary
requests